react-table: move helpers file into subdirectory
