+++ app/assets/javascripts/Components/Helpers/table_helpers.jsx
