+++ app/assets/javascripts/Components/ta_table.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {render} from 'react-dom';
 
 import ReactTable from 'react-table';
-import {stringFilter} from './table_helpers';
+import {stringFilter} from './Helpers/table_helpers';
 import {CheckboxTable} from "./markus_with_selection_hoc";
